Move `html_diff_render` into its own module
Also moves its “validity” tests (the ones that can actually fail) into their own module.
